test: move get FF and envs to globalTeardown
