+++ apps/ledger-live-mobile/e2e/specs/onboarding.spec.ts
@@ -1,6 +1,5 @@
 import { device } from "detox";
-import { isAndroid } from "../helpers";
-import { launchApp } from "../setup";
+import { isAndroid, launchApp } from "../helpers";
 import { Application } from "../page";
 
 let app: Application;
